Edit view

+++ database/migrations/2018_06_25_000000_create_users_table.php
@@ -20,9 +20,10 @@ class CreateUsersTable extends Migration
             $table->string('email')->nullable();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password')->nullable();
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->rememberToken();
             $table->timestamps();
+
             $table->foreign('role_id')->references('id')->on('roles');
         });
     }
