+++ database/migrations/2018_08_19_120111_create_mades_table.php
@@ -16,7 +16,7 @@ class CreateMadesTable extends Migration
         Schema::create('mades', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name');
-            $table->integer('status');
+            $table->bigInteger('status');
             $table->timestamps();
         });
     }
