+++ database/migrations/2018_08_19_120114_create_subscribes_table.php
@@ -16,7 +16,7 @@ class CreateSubscribesTable extends Migration
         Schema::create('subscribes', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('email');
-            $table->integer('status');
+            $table->bigInteger('status');
             $table->timestamps();
         });
     }
