+++ database/migrations/2018_09_03_112511_create_purchasedgoods_table.php
@@ -15,8 +15,8 @@ class CreatePurchasedgoodsTable extends Migration
     {
         Schema::create('purchasedgoods', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('product_id');
             $table->foreign('user_id')->references('id')->on('users');
             $table->foreign('product_id')->references('id')->on('products');
             $table->timestamps();
