+++ database/migrations/2019_03_30_145720_create_cache_table.php
@@ -16,7 +16,7 @@ class CreateCacheTable extends Migration
         Schema::create('cache', function (Blueprint $table) {
             $table->string('key')->unique();
             $table->mediumText('value');
-            $table->integer('expiration');
+            $table->bigInteger('expiration');
         });
     }
